Refactor package structure as domain structure
