+++ src/main/java/com/example/baemin/domain/category/CategoryService.java
@@ -1,8 +1,5 @@
-package com.example.baemin.domain.category.service;
+package com.example.baemin.domain.category;
 
-import com.example.baemin.domain.category.dto.CategoryDTO;
-import com.example.baemin.domain.category.repository.CategoryRepository;
-import com.example.baemin.domain.entity.Category;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
