+++ src/main/java/com/example/baemin/domain/common/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.example.baemin.domain.entity;
+package com.example.baemin.domain.common;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 
 @EntityListeners(value = {AuditingEntityListener.class})
 @MappedSuperclass
-public abstract class Base {
+public abstract class BaseEntity {
     @CreatedDate
     @Column(updatable = false)
     private LocalDateTime createdAt;
