+++ src/main/java/com/example/baemin/domain/restaurant/Restaurant.java
@@ -1,5 +1,7 @@
-package com.example.baemin.domain.entity;
+package com.example.baemin.domain.restaurant;
 
+import com.example.baemin.domain.common.BaseEntity;
+import com.example.baemin.domain.food.Food;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,16 +13,19 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import org.hibernate.annotations.ColumnDefault;
+import org.hibernate.annotations.DynamicInsert;
 
 import java.util.List;
 
 @Entity
+@DynamicInsert
 @Getter
 @Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class Category extends Base {
+public class Restaurant extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -28,6 +33,13 @@ public class Category extends Base {
     @Column(nullable = false)
     private String name;
 
-    @OneToMany(mappedBy = "category")
+    @Column(nullable = false)
+    private String location;
+
+    @ColumnDefault("true")
+    private Boolean isActive;
+
+    @OneToMany(mappedBy = "restaurant")
     private List<Food> food;
+
 }
