+++ src/main/resources/application.yml
@@ -1,3 +1,6 @@
+#server:
+#  servlet:
+#    contextPath: /api
 spring:
   datasource:
     driver-class-name: org.mariadb.jdbc.Driver
